+++ src/components/Form/index.ts
@@ -1,6 +1,7 @@
 import { Form as BaseForm } from "./Form";
 import { Control } from "./FormControl";
 import { Description } from "./FormDescription";
+import { EmailDuplicateCheckButton } from "./FormEmailDuplicateCheckButton";
 import { Field } from "./FormField";
 import { ImageUpload } from "./FormImageUpload";
 import { Input } from "./FormInput";
@@ -8,6 +9,7 @@ import { Item } from "./FormItem";
 import { Label } from "./FormLabel";
 import { LoginLink } from "./FormLoginLink";
 import { Message } from "./FormMessage";
+import { NicknameDuplicateCheckButton } from "./FormNicknameDuplicateCheckButton";
 import { Radio } from "./FormRadio";
 import { Select } from "./FormSelect";
 import { SignupLink } from "./FormSignupLink";
@@ -31,6 +33,8 @@ export const Form = Object.assign(BaseForm, {
   Select,
   ImageUpload,
   Radio,
+  EmailDuplicateCheckButton,
+  NicknameDuplicateCheckButton,
 });
 
 export * from "./types";
